- UART - Finished revamp of basic UART library functionality. It now utilizes the private object more and has more abstraction away from the hardware. It allows the user to choose a major and minor mode which then sets up the hardware in specific ways. The buffer system has also been updated and cleaned up. Each direction (TX and RX) have four buffer options, hardware FIFO only, DMA, software, and hybrid (i.e. DMA+software). The user can choose the type of buffer and the size of the buffer for each type.
- Delays -
Added some delay functions to the library.

- DMA -
Cleaned up the DMA channel library. Made it a bit closer to the hardware so that it can be a bit more flexible in terms of how it is set up.

- Bit Ops -
Added some new bit operations and fixed a few bugs. The macros need to be renamed, however, because they have a terrible naming convention.